backend/requirements.txt
fastapi[standard]
uvicorn[standard]
langgraph
langchain-openai
langchain-core
langchain
pydantic>=2.0
aiosqlite
httpx
python-dotenv
sse-starlette
